+++ NEWS.md
@@ -1,7 +1,10 @@
 ## 0.5.0 to 0.5.1 (unreleased so far)
 
  (TODO)
- * Contactdb: Updates fixed list of common "tags".
+ * Contactdb: The list of common "tags" can be configured now.
+
+### Upgrade
+ * Configuration: (optional) Add "common\_tags" to the contactdb config file.
 
 ## 0.4.3.dev0 to 0.5.0
  * Contactdb: Adds example script for importing a manual whitelist.